requirements.txt
MLB-StatsAPI==1.2
google-cloud-pubsub==2.5.0
google-cloud-storage==1.38.0
google-cloud-secret-manager==2.5.0
pandas==1.2.4